Refactor item schema and implementation to separate unit from price
The schema for an item used to combine both price and unit in its price
field as a string. This was a tedious problem as string concatenatation
and splitting was needed upon access for these values. Price and unit
will now be accessible as separate fields in the item schema.