Implement image schema in item schema
Images for an item used to be stored in the item's "images" subdocument
by their cloudinary public IDs. This was a problem because in order to
access the image URLs, an additional fetch request was necessary. Each
image's public ID and URL will now be stored in an image schema, and
image URLs will now be accessed immediately.